base.mk: towards variants
Actually support some of the flags that we promise, rip out dep file
support as it's currently broken.

XXX: removing the $(call debug) in some of the macros causes make to
break with "multiple target patterns". Cause is unclear. $(call debug)
was added to debug it and then cause the issue to vanish. Might be some
interaction that causes joining of lines across macros.